Initial hack to parse St George PDF statements
Motivated by a change in policy which means I cannot download QIF files for
transactions older than my previous statement, and not wanting to hand-type
a year's worth of transactions into my accounting software.

So far it just extracts all the information from the three statements I happen
to have around, with a lot of guessing and assertions to check the guessing.